ais/ais-backend.new/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]
passlib==1.7.4
pyjwt
pydantic[email]
python-jose[cryptography]
python-multipart
jose

